My current project
So far I have created the front end and the back end of my project. The
front end contains the jsp, html, css, js and scss files along with a
few java files like the controllers.

In the back end project, it contains the Dao's , the DAO
implementations, the models and the configuration files for hibernate.
Along with these files, there are test files to check whether the back
end is connected to the H2 database.